requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
psutil==5.9.6
psycopg2-binary==2.9.7
boto3==1.29.7
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0
sqlalchemy==2.0.21